Rename OrdinoAppComponent to OrdinoApp
