+++ src/internal/components/session/CurrentSessionCard.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {Button, Dropdown} from 'react-bootstrap';
 import {ProvenanceGraphMenuUtils} from 'tdp_core';
 import {IStartMenuSessionSectionDesc} from '../../..';
-import {GraphContext} from '../../OrdinoAppComponent';
+import {GraphContext} from '../../OrdinoApp';
 import {ListItemDropdown} from '../common';
 import {EAction, CommonSessionCard} from './CommonSessionCard';
 import {SessionListItem} from './SessionListItem';
