+++ src/internal/components/session/SavedSessionCard.tsx
@@ -4,7 +4,7 @@ import {Tab, Nav, Row, Col, Button, Dropdown} from 'react-bootstrap';
 import {ProvenanceGraphMenuUtils} from 'tdp_core';
 import {IStartMenuSessionSectionDesc} from '../../..';
 import {useAsync} from '../../../hooks';
-import {GraphContext} from '../../OrdinoAppComponent';
+import {GraphContext} from '../../OrdinoApp';
 import {ListItemDropdown} from '../common';
 import {EAction, CommonSessionCard} from './CommonSessionCard';
 import {SessionListItem} from './SessionListItem';
