+++ src/internal/index.ts
@@ -2,7 +2,7 @@ export * from './constants';
 export * from './ViewWrapper';
 export * from './IOrdinoApp';
 export * from './cmds';
-export * from './OrdinoAppComponent';
+export * from './OrdinoApp';
 export * from './StartMenu';
 export * from './menu';
 export * from './components';
